+++ changelog.md
@@ -1,2 +1,2 @@
 - Updated to 26.2
-- Updated Belarusian (Cyrillic) translation
+- Added configurable amount of items and columns in the bundle tooltip